tests/wpt/web-platform-tests/web-animations/interfaces/AnimationEffectTiming/iterationStart.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffectTiming.iterationStart</title>
<link rel="help" href="https://drafts.csswg.org/web-animations/#dom-animationeffecttiming-iterationstart">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(t => {
  const anim = createDiv(t).animate(null);
  assert_equals(anim.effect.timing.iterationStart, 0);
}, 'Has the default value 0');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] },
                           { iterationStart: 0.2,
                             iterations: 1,
                             fill: 'both',
                             duration: 100,
                             delay: 1 });
  anim.effect.timing.iterationStart = 2.5;
  assert_times_equal(anim.effect.getComputedTiming().progress, 0.5);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 2);
}, 'Changing the value updates computed timing when backwards-filling');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] },
                           { iterationStart: 0.2,
                             iterations: 1,
                             fill: 'both',
                             duration: 100,
                             delay: 0 });
  anim.effect.timing.iterationStart = 2.5;
  assert_times_equal(anim.effect.getComputedTiming().progress, 0.5);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 2);
}, 'Changing the value updates computed timing during the active phase');

test(t => {
  const div = createDiv(t);
  const anim = div.animate({ opacity: [ 0, 1 ] },
                           { iterationStart: 0.2,
                             iterations: 1,
                             fill: 'both',
                             duration: 100,
                             delay: 0 });
  anim.finish();
  anim.effect.timing.iterationStart = 2.5;
  assert_times_equal(anim.effect.getComputedTiming().progress, 0.5);
  assert_equals(anim.effect.getComputedTiming().currentIteration, 3);
}, 'Changing the value updates computed timing when forwards-filling');

test(t => {
  const div = createDiv(t);
  const anim = div.animate(null);
  for (let invalid of [-1, NaN, Infinity]) {
    assert_throws({ name: 'TypeError' }, () => {
      anim.effect.timing.iterationStart = invalid;
    }, `setting ${invalid}`);
    assert_throws({ name: 'TypeError' }, () => {
      div.animate({}, { iterationStart: invalid });
    }, `animate() with ${invalid}`);
  }
}, 'Throws when setting invalid values');

</script>
</body>
